requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
mysql-connector==2.2.9
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
